+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 1.1.0
 * Improved ppm labels for ABfit plot results.
 * Updated unit tests.
+* Added vignette on manually adjusting ABfit baseline smoothness.
 
 # spant 1.0.0
 * ABfit analysis method has been added, and is now default for the mrs_fit